Added initial folder structure
