+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "mobile-dev",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
