+++ src/components/molecules/.gitkeep
@@ -1,0 +1,2 @@
+# Specific components built with atoms/
+# may or may not be reusable
