+++ src/screens/.gitignore
@@ -1,0 +1,2 @@
+# High level pages/screens/components/
+# Not intended for reusability
